Revert "moved presence, inprog"
This reverts commit 02820b267ebb76e7af1f1fee55ebedc5de164c3d.

revert inprog
